server/requirements.txt
flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.0
supabase==2.3.5
google-generativeai==0.8.5
pytest
pytest-mock 